invert_limb macro -> n_preinvert_limb_prenorm; de-inline
